meusvideos-service/requirements.txt
django>=3.0,<4.0
djangorestframework>=3.0,<4.0
django-filter>=2.0,<3.0
dj-rest-auth>=1.0,<2
djangorestframework-simplejwt>=4.0,<5.0
django-allauth>=0.44.0,<1.0
